Add confirm dialog for deletion

+++ src/app/modules/annotation-schema/code-smell-detail/code-smell-detail.component.ts
@@ -3,7 +3,9 @@ import { MatDialog } from "@angular/material/dialog";
 import { MatTable, MatTableDataSource } from "@angular/material/table";
 import { ActivatedRoute, Params } from "@angular/router";
 import { ToastrService } from "ngx-toastr";
+import { DialogConfigService } from "../../data-set/dialogs/dialog-config.service";
 import { AddHeuristicDialogComponent } from "../dialogs/add-heuristic-dialog/add-heuristic-dialog.component";
+import { ConfirmDialogComponent } from "../dialogs/confirm-dialog/confirm-dialog.component";
 import { UpdateHeuristicDialogComponent } from "../dialogs/update-heuristic-dialog/update-heuristic-dialog.component";
 import { CodeSmellDefinition } from "../model/code-smell-definition/code-smell-definition.model";
 import { Heuristic } from "../model/heuristic/heuristic.model";
@@ -71,9 +73,16 @@ export class CodeSmellDetailComponent implements OnInit {
   }
 
   public removeHeuristic(heuristic: Heuristic): void {
-    this.heuristicsDataSource.data.splice(this.heuristicsDataSource.data.findIndex(h => h.id == heuristic.id), 1);
-    this.annotationSchemaService.removeHeuristicFromCodeSmell(this.chosenCodeSmell.id, heuristic.id)
-    .subscribe(res => this.table.renderRows());
+    let dialogConfig = DialogConfigService.setDialogConfig('auto', 'auto', 'Deleted heuristic will be removed from the existing annotations.');
+    let dialogRef = this.dialog.open(ConfirmDialogComponent, dialogConfig);
+    dialogRef.updateSize('20%');
+    dialogRef.afterClosed().subscribe((confirmed: boolean) => {
+      if (confirmed) {
+        this.heuristicsDataSource.data.splice(this.heuristicsDataSource.data.findIndex(h => h.id == heuristic.id), 1);
+        this.annotationSchemaService.removeHeuristicFromCodeSmell(this.chosenCodeSmell.id, heuristic.id)
+          .subscribe(() => this.table.renderRows());
+      }
+    });
   }
 
   public updateHeuristic(heuristic: Heuristic): void {
